+++ src/jna-client/org/firebirdsql/gds/ng/jna/JnaDatabase.java
@@ -25,6 +25,7 @@ import org.firebirdsql.gds.*;
 import org.firebirdsql.gds.impl.DatabaseParameterBufferExtension;
 import org.firebirdsql.gds.ng.*;
 import org.firebirdsql.gds.ng.listeners.TransactionListener;
+import org.firebirdsql.jdbc.FBDriverNotCapableException;
 import org.firebirdsql.jna.fbclient.FbClientLibrary;
 import org.firebirdsql.jna.fbclient.ISC_STATUS;
 import org.firebirdsql.jna.fbclient.WinFbClientLibrary;
@@ -365,6 +366,18 @@ public class JnaDatabase extends AbstractFbDatabase<JnaDatabaseConnection>
     @Override
     public int getHandle() {
         return handle.getValue();
+    }
+
+    @Override
+    public void setNetworkTimeout(int milliseconds) throws SQLException {
+        throw new FBDriverNotCapableException(
+                "Setting network timeout not supported in native implementation");
+    }
+
+    @Override
+    public int getNetworkTimeout() throws SQLException {
+        throw new FBDriverNotCapableException(
+                "Getting network timeout not supported in native implementation");
     }
 
     public IntByReference getJnaHandle() {