+++ src/main/org/firebirdsql/gds/JaybirdErrorCodes.java
@@ -82,6 +82,8 @@ public interface JaybirdErrorCodes {
     int jb_explainedExecutionPlanNotSupported            = 337248293;
     int jb_failedToLoadNativeLibrary                     = 337248294;
     int jb_blobClosed               = 337248295;
+    int jb_invalidTimeout           = 337248296;
+    int jb_invalidExecutor          = 337248297;
 
     @SuppressWarnings("unused")
     int jb_range_end                = 337264639;