+++ src/main/org/firebirdsql/gds/ng/wire/FbWireOperations.java
@@ -211,6 +211,17 @@ public interface FbWireOperations {
             ProcessAttachCallback processAttachCallback)
             throws IOException, SQLException;
 
+    /**
+     * Sets the network timeout for this attachment.
+     *
+     * @param milliseconds
+     *         Timeout in milliseconds; 0 means no timeout. If the attachment doesn't support milliseconds, it should
+     *         round up to the nearest second.
+     * @throws SQLException If this attachment is closed, the value of {@code milliseconds} is smaller than 0, or if setting the timeout fails.
+     * @throws java.sql.SQLFeatureNotSupportedException If this attachment doesn't support changing the network timeout.
+     */
+    void setNetworkTimeout(int milliseconds) throws SQLException;
+
     interface ProcessAttachCallback {
         void processAttachResponse(GenericResponse response);
     }